Add custom exceptions for baked goods and purchases; refactor purchase model to use Decimal

+++ models/purchase.py
@@ -1,5 +1,6 @@
 from dataclasses import dataclass
 import datetime
+from decimal import Decimal
 from numbers import Number
 from models.customer import Customer, Ingredient
 from models.drink import Drink
@@ -10,5 +11,5 @@ class Purchase:
     id: Number
     timestamp: datetime
     items: list[Drink, Baked_good]
-    total_cost: float
+    total_cost: Decimal
     Customer: Customer
